0.8
fixed some stuff that Jayden added. made it so that there is a public variable "weapon damage" which allows us to control the base damage of the weapon instead of hardcoding it to 40. We should also make the accuracy, total damage dealt, etc that he added into stats in the PlayerStats class because they aren't gun specific and each instance of the  Weapon class would be for a specific gun.

We only have one Weapon for now so I'm gonna leave that latter part as is.