0.9
stats on the players accuracy, total damage dealt, total deaths, total hits, total shots, total kills, and total deaths are now stored in the GameMaster. This keeps them weapon agnostic and when the player respawns. I also moved a lot of stuff out of the update that didn't need to be there. for example we do not need to calculate accuracy every frame, only when a shot is fired.