2.0
isDead boolean is now in the Enemy class instead of the specific AI class. working on second type of enemy; a Korean Jindo with dynamite strapped to it's back. I think we need a smaller sprite for this character; the dog is YUGE right now, let me tell you. It runs after trump and suicides itself when it gets near him. I am calculating distance between other enemies and trump upon suicide, but not applying damage yet. More to come.